Make settings mutable by default and remote "update_settings()"
Updating/monkeypatching settings in tests is very painful when settings
are immutable and especially nested settings classes are used.

Furthermore, there may be legitimate reasons for updating the settings
while a program is running.

For theses reasons, settings are now mutable by default but can still
be made immutable.  Settings loading is not affected by this.

The "update_settings()" function was also very clumsy because it could
only change a single value with each call.  Using "attr.evolve()" is
much more convenient.